minSdkVersion = 15, refactored compatibility classes to match
