+++ AnkiDroid/src/main/java/com/ichi2/compat/Compat.java
@@ -17,9 +17,7 @@
 package com.ichi2.compat;
 
 import android.content.Context;
-import android.content.Intent;
 import android.net.Uri;
-import android.speech.tts.TextToSpeech;
 import android.view.View;
 import android.view.Window;
 import android.webkit.WebSettings;
@@ -27,7 +25,6 @@ import android.widget.RemoteViews;
 
 import com.ichi2.anki.AbstractFlashcardViewer;
 import com.ichi2.anki.AnkiActivity;
-import com.ichi2.anki.DeckPicker;
 
 import java.io.File;
 
@@ -39,7 +36,7 @@ import io.requery.android.database.sqlite.SQLiteDatabase;
  * A set of implementations for the supported platforms are available.
  * <p>
  * Each implementation ends with a {@code V<n>} prefix, identifying the minimum API version on which this implementation
- * can be used. For example, see {@link CompatV10}.
+ * can be used. For example, see {@link CompatV15}.
  * <p>
  * Each implementation should extend the previous implementation and implement this interface.
  * <p>
@@ -48,28 +45,23 @@ import io.requery.android.database.sqlite.SQLiteDatabase;
  * Methods not supported by its API will default to the empty implementations of CompatV8.  Methods first supported
  * by lower APIs will default to those implementations since we extended them.
  * <p>
- * Example: CompatV9 extends CompatV8. This means that the nfcNormalized function that uses classes only available
- * in API 9, should be implemented properly in CompatV9 with @Override annotation. On the other hand a method
- * like setOverScrollModeNever that first becomes available in API 9 need not be implemented again in CompatV10,
+ * Example: CompatV21 extends CompatV19. This means that the setSelectableBackground function using APIs only available
+ * in API 21, should be implemented properly in CompatV19 with @Override annotation. On the other hand a method
+ * like showViewWithAnimation that first becomes available in API 19 need not be implemented again in CompatV21,
  * unless the behaviour is supposed to be different there.
  */
 public interface Compat {
 
     /* Mock the Intent PROCESS_TEXT constants introduced in API 23. */
-    public static final String ACTION_PROCESS_TEXT = "android.intent.action.PROCESS_TEXT";
-    public static final String EXTRA_PROCESS_TEXT = "android.intent.extra.PROCESS_TEXT";
+    String ACTION_PROCESS_TEXT = "android.intent.action.PROCESS_TEXT";
+    String EXTRA_PROCESS_TEXT = "android.intent.extra.PROCESS_TEXT";
 
     String detagged(String txt);
-    void setTtsOnUtteranceProgressListener(TextToSpeech tts);
     void disableDatabaseWriteAheadLogging(SQLiteDatabase db);
-    void enableCookiesForFileSchemePages();
     void updateWidgetDimensions(Context context, RemoteViews updateViews, Class<?> cls);
-    void restartActivityInvalidateBackstack(AnkiActivity activity);
     void setFullScreen(AbstractFlashcardViewer activity);
     void setSelectableBackground(View view);
     void openUrl(AnkiActivity activity, Uri uri);
-    void supportAddContentMenu(final DeckPicker a);
-    Intent getPreferenceSubscreenIntent(Context context, String subscreen);
     void prepareWebViewCookies(Context context);
     void flushWebViewCookies();
     void setHTML5MediaAutoPlay(WebSettings settings, Boolean allow);
